fix: nil ref object getting initialized
This patch fixes a VERY long standing issue related to ref object being
initialized while they were actually stored as nil after going to
bson.Unmarshal.

The reason is that we must check if the raw.Kind is bson.ElementNil and
if not we must return bson.ErrSetZero.

I found this while crusing the mgo code and it was definitely not
obvious