elegen: call elemental.ResetDefaultForZeroValue() on Validate()
Previously, ResetDefaultForZeroValue was only call on ref, refList etc
in the validate function, assuming something would call it before.

This seem to be an actual very long running issue where zero values were
not reseted anywhere in the code, and was simply assuming the client SDK
would do it.

The entire idea to not call this on Validate, or during the write
operation (right now it was only call on read operation) was to have the
default evolve with the spec if it has never been set. This has never
worked as the data was either init'ed to the default by the client SDK,
or it would trigger a 422 required prop error anyways.

Better solidfy this logic here now.